pyproject.toml
[project]
name = "url-lookup-service"
version = "0.1.0"
description = "URL lookup service for malware detection"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
